few more modifications
Changed hard coding of DB in core class.
Moved index page story and info to a table. Data is defined in the
data.inc.php file and imported with the deploy script.
Most of the deploy script is done. Need to add in the details page info
to be inserted into a table.