Stupid bugs and additions
Fixed a bugo n the guestpage, where it was using $this-> instead of $wedding->

Added the RSVP send from and to email properties. So the RSVP and Guest Book can have a separate email if need be.

Fixed the error_log warnings that were popping up about undefined offset or index.

Fixed a smarty issue in the rsvp.php page where I was not defining a variable like I was supposed to.