Few more changes
Fixed a bug with the RSVP reservation not sending a user back to the
start if they are missing required information.

Added a basic spam filter for the guestbook.

Guestbook will not also format the messages to be a set width now so
that the view page looks good.

renamed sample config and data files to no longer have sample in the
filename. Makes it easier to use to be deployed.