Added word wrapping for RSVP error messages
Error messages for the RSVP form will be wrapped to 50 chars. They also
were set to be on top of the corner images so the text is not hidden any
more.

Also added the class properties to the definition part and set them as
public that are neccessary.

Also added an environment variable that can be set in the config.inc.php
called $env. This is a string that will be used to denote what
environment the emails and messages are coming from in the subject.